Update NASA_app.py
program sends GET request to NASA DONKI API for CME data.  Program prints a list of CMEs within last 30 days that either have or are predicted to impact Earth.  
PROBLEM: as it is currently written, the program will simply create an empty text file the first time it is run.  Once the text file is created, the program works fine.