Create CME_updater.py
this program uses Tkinter to create a UI window.  The user has the option to view a list of new coronal mass ejection (CME) events identified in the last 30 days.  This list is generated via a GET request to NASA's DONKI CME API.  The user also has the option to save this list to a text file.  This text file will be generated in whatever folder the program is hosted.